+++ ntex-codec/src/framed.rs
@@ -3,6 +3,7 @@ use std::task::{Context, Poll};
 use std::{fmt, io};
 
 use bytes::{Buf, BytesMut};
+use either::Either;
 use futures_core::{ready, Stream};
 use futures_sink::Sink;
 
@@ -250,7 +251,7 @@ where
     }
 
     /// Flush write buffer to underlying I/O stream.
-    pub fn flush(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), U::Error>> {
+    pub fn flush(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), io::Error>> {
         log::trace!("flushing framed transport");
 
         let len = self.write_buf.len();
@@ -269,8 +270,7 @@ where
                         return Poll::Ready(Err(io::Error::new(
                             io::ErrorKind::WriteZero,
                             "failed to write frame to transport",
-                        )
-                        .into()));
+                        )));
                     } else {
                         written += n
                     }
@@ -278,7 +278,7 @@ where
                 Poll::Ready(Err(e)) => {
                     log::trace!("Error during flush: {}", e);
                     self.flags.insert(Flags::DISCONNECTED);
-                    return Poll::Ready(Err(e.into()));
+                    return Poll::Ready(Err(e));
                 }
             }
         }
@@ -348,7 +348,7 @@ where
     pub fn next_item(
         &mut self,
         cx: &mut Context<'_>,
-    ) -> Poll<Option<Result<U::Item, U::Error>>> {
+    ) -> Poll<Option<Result<U::Item, Either<U::Error, io::Error>>>> {
         let mut done_read = false;
 
         loop {
@@ -364,12 +364,12 @@ where
                         Ok(Some(frame)) => Poll::Ready(Some(Ok(frame))),
                         Ok(None) => {
                             if let Some(err) = self.err.take() {
-                                Poll::Ready(Some(Err(err.into())))
+                                Poll::Ready(Some(Err(Either::Right(err))))
                             } else {
                                 Poll::Ready(None)
                             }
                         }
-                        Err(e) => return Poll::Ready(Some(Err(e))),
+                        Err(e) => return Poll::Ready(Some(Err(Either::Left(e)))),
                     };
                 }
 
@@ -380,7 +380,7 @@ where
                         log::trace!("frame decoded from buffer");
                         return Poll::Ready(Some(Ok(frame)));
                     }
-                    Err(e) => return Poll::Ready(Some(Err(e))),
+                    Err(e) => return Poll::Ready(Some(Err(Either::Left(e)))),
                     _ => (), // Need more data
                 }
 
@@ -428,7 +428,7 @@ where
                             self.flags.insert(Flags::EOF | Flags::READABLE);
                             break;
                         } else {
-                            return Poll::Ready(Some(Err(e.into())));
+                            return Poll::Ready(Some(Err(Either::Right(e))));
                         }
                     }
                 }
@@ -442,7 +442,7 @@ where
     T: AsyncRead + Unpin,
     U: Decoder + Unpin,
 {
-    type Item = Result<U::Item, U::Error>;
+    type Item = Result<U::Item, Either<U::Error, io::Error>>;
 
     #[inline]
     fn poll_next(
@@ -486,7 +486,7 @@ where
         mut self: Pin<&mut Self>,
         cx: &mut Context<'_>,
     ) -> Poll<Result<(), Self::Error>> {
-        self.flush(cx)
+        self.flush(cx).map_err(From::from)
     }
 
     #[inline]